+++ reporter/csv.go
@@ -27,7 +27,7 @@ type CSV struct {
 // Initialize the system
 func (s *CSV) Initialize(w *ecs.World) {
 	s.Observer.Initialize(w)
-	s.header = s.Observer.Header(w)
+	s.header = s.Observer.Header()
 	if s.Sep == "" {
 		s.Sep = ","
 	}